+++ XYUI/src/XYUI.Avalonia/Foundation/XYDensity.cs
@@ -1,0 +1,8 @@
+namespace XYUI.Avalonia.Foundation;
+
+public enum XYDensity
+{
+    Compact,
+    Default,
+    Comfortable
+}